refactor: extract shared ring painter, page-dots indicator, and card shell to remove duplication
